refactor backend code to use services, move custom hooks to separate file, fix bug, add @ to import statements
