+++ src/models/error/HttpError.tsx
@@ -1,0 +1,8 @@
+export class HttpError extends Error {
+    statusCode: number;
+
+    constructor(message: string, statusCode: number) {
+      super(message);
+      this.statusCode = statusCode;
+    }
+  }
